chore: masking exception user table
