+++ frontend/src/components/ColumnDataTable/index.vue
@@ -113,6 +113,7 @@ const showSensitiveColumn = computed(() => {
       engine.value === Engine.ORACLE ||
       engine.value === Engine.SNOWFLAKE ||
       engine.value === Engine.MSSQL ||
+      engine.value === Engine.BIGQUERY ||
       engine.value === Engine.RISINGWAVE)
   );
 });
