+++ frontend/src/types/member.ts
@@ -1,5 +1,6 @@
 // IAM related resource definition.
 export interface DatabaseResource {
+  instanceResourceId?: string;
   // the database full name
   databaseName: string;
   schema?: string;
